Use python build script
